backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pillow==11.3.0
pydantic-settings==2.11.0
huggingface_hub[cli]==0.34.4
torch==2.4.1
diffusers==0.34.0
transformers==4.55.4
accelerate==1.10.0
safetensors==0.6.2
git+https://github.com/TencentARC/PhotoMaker.git
